+++ tomlkit/__init__.py
@@ -25,7 +25,7 @@ from .api import value
 from .api import ws
 
 
-__version__ = "0.8.0"
+__version__ = "0.9.0"
 __all__ = [
     "aot",
     "array",