+++ CHANGELOG.md
@@ -2,10 +2,16 @@
 
 ## [Unreleased]
 
+## [0.9.0] - 2022-02-01
+
 ### Added
 
 - Add a new argument to `table` API to allow it to be a super table. ([#159](https://github.com/sdispater/tomlkit/pull/159))
 - Support adding item to `Table` and `Container` with dotted key. ([#160](https://github.com/sdispater/tomlkit/pull/160))
+
+### Fixed
+
+- Fix a bug of `value()` API that parses string incompletely. ([#168](https://github.com/sdispater/tomlkit/pull/168))
 
 ## [0.8.0] - 2021-12-20
 
@@ -228,7 +234,8 @@
 - Fixed handling of super tables with different sections.
 - Fixed raw strings escaping.
 
-[unreleased]: https://github.com/sdispater/tomlkit/compare/0.8.0...master
+[unreleased]: https://github.com/sdispater/tomlkit/compare/0.9.0...master
+[0.9.0]: https://github.com/sdispater/tomlkit/releases/tag/0.9.0
 [0.8.0]: https://github.com/sdispater/tomlkit/releases/tag/0.8.0
 [0.7.2]: https://github.com/sdispater/tomlkit/releases/tag/0.7.2
 [0.7.1]: https://github.com/sdispater/tomlkit/releases/tag/0.7.1